requirements.txt
discord.py==2.3.2
Pillow==10.0.1
python-dotenv==1.0.0
aiohttp==3.8.6
requests==2.31.0 